requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
django-tempus-dominus==5.1.2.13
django-widget-tweaks==1.4.8
Pillow==7.2.0
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
